some minor changes and improvement
