+++ helix-core/src/main/java/org/apache/helix/controller/rebalancer/waged/model/ClusterModelProvider.java
@@ -32,6 +32,7 @@ import java.util.stream.Collectors;
 import org.apache.helix.HelixConstants;
 import org.apache.helix.HelixException;
 import org.apache.helix.controller.dataproviders.ResourceControllerDataProvider;
+import org.apache.helix.controller.rebalancer.util.DelayedRebalanceUtil;
 import org.apache.helix.model.ClusterConfig;
 import org.apache.helix.model.IdealState;
 import org.apache.helix.model.InstanceConfig;
@@ -60,21 +61,27 @@ public class ClusterModelProvider {
   }
 
   /**
+   * TODO: On integration with WAGED,
+   *   we need to combine current state with computed assignment based on the cluster, the partitions on delayed
+   *   instances have to be allocated. This should happen after the algorithm computation.
+   *   Also integrate with counter and latency metrics -- qqu
    * Compute a new Cluster Model with scope limited to partitions with best possible assignment missing minActiveReplicas
    * because of delayed rebalance setting.
    * @param dataProvider The controller's data cache
    * @param resourceMap The full map of the resource by name
    * @param activeInstances The active instances that will be used in the calculation.
-   * @param bestPossibleAssignment The persisted Best Possible assignment that was generated in the previous rebalance.
+   * @param resourceAssignment The resource assignment state to compute on. This should be the current state assignment;
+   *                           if it's run right after another rebalance calculation, the best possible assignment from
+   *                           previous result can be used.
    * @return the ClusterModel
    */
   public static ClusterModel generateClusterModelForDelayedRebalanceOverwrites(
       ResourceControllerDataProvider dataProvider,
       Map<String, Resource> resourceMap,
       Set<String> activeInstances,
-      Map<String, ResourceAssignment> bestPossibleAssignment) {
+      Map<String, ResourceAssignment> resourceAssignment) {
     return generateClusterModel(dataProvider, resourceMap, activeInstances, Collections.emptyMap(),
-        Collections.emptyMap(), bestPossibleAssignment, RebalanceScopeType.DELAYED_REBALANCE_OVERWRITES);
+        Collections.emptyMap(), resourceAssignment, RebalanceScopeType.DELAYED_REBALANCE_OVERWRITES);
   }
 
   /**
@@ -216,7 +223,7 @@ public class ClusterModelProvider {
         break;
       case DELAYED_REBALANCE_OVERWRITES:
         toBeAssignedReplicas =
-            DelayedRebalanceOverwriteUtil.findToBeAssignedReplicasForMinActiveReplica(dataProvider, replicaMap, activeInstances,
+            DelayedRebalanceUtil.findToBeAssignedReplicasForMinActiveReplica(dataProvider, replicaMap, activeInstances,
                 currentAssignment, allocatedReplicas);
         break;
       default:
@@ -500,7 +507,7 @@ public class ClusterModelProvider {
   }
 
   // <partition, <state, instances set>>
-  static Map<String, Map<String, Set<String>>> getStateInstanceMap(
+  public static Map<String, Map<String, Set<String>>> getStateInstanceMap(
       ResourceAssignment assignment) {
     if (assignment == null) {
       return Collections.emptyMap();
@@ -580,16 +587,6 @@ public class ClusterModelProvider {
       }
       return new HashMap.SimpleEntry<>(resourceName, replicas);
     }).collect(Collectors.toMap(entry -> entry.getKey(), entry -> entry.getValue()));
-  }
-
-  private static Map<String, Map<String, List<AssignableReplica>>> getAssignableReplicasByResourcePartition(
-      Map<String, Set<AssignableReplica>> replicaMap) {
-    return replicaMap.entrySet()
-        .stream()
-        .collect(Collectors.toMap(
-            Map.Entry::getKey,
-            e -> e.getValue().stream().collect(Collectors.groupingBy(AssignableReplica::getPartitionName)))
-        );
   }
 
   /**
